// frontend/src/models/student-summary.ts
// Generated by https://quicktype.io

export interface IStudentSummary {
  status: string;
  data: Data;
}

export interface Data {
  reports: Report[];
}

export interface Report {
  _id: ID;
  numberOfAbsences: number;
  Akceptovaných: number;
  Zamietnutých: number;
}

export interface ID {
  course: string;
  courseType: string;
}